Move users to separate module
